# Conflicts:
#	backend/molgenis-emx2-graphql/src/main/java/org/molgenis/emx2/graphql/GraphqlAdminFieldFactory.java
#	backend/molgenis-emx2-graphql/src/main/java/org/molgenis/emx2/graphql/GraphqlApiFactory.java

+++ backend/molgenis-emx2-sql/src/main/java/org/molgenis/emx2/sql/SqlDatabase.java
@@ -4,6 +4,7 @@ import static org.jooq.impl.DSL.name;
 import static org.molgenis.emx2.ColumnType.STRING;
 import static org.molgenis.emx2.Constants.MG_USER_PREFIX;
 import static org.molgenis.emx2.Constants.SYSTEM_SCHEMA;
+import static org.molgenis.emx2.sql.MetadataUtils.*;
 import static org.molgenis.emx2.sql.SqlDatabaseExecutor.*;
 import static org.molgenis.emx2.sql.SqlSchemaMetadataExecutor.executeCreateSchema;
 
@@ -471,11 +472,44 @@ public class SqlDatabase extends HasSettings<Database> implements Database {
   @Override
   public void removeUser(String user) {
     long start = System.currentTimeMillis();
+    if (user.equals("admin")) throw new MolgenisException("You cant remove admin");
+    if (user.equals("anonymous")) throw new MolgenisException("You cant remove anonymous");
+
     if (!hasUser(user))
       throw new MolgenisException(
           "Remove user failed: User with name '" + user + "' doesn't exist");
     tx(db -> ((SqlDatabase) db).getJooq().execute("DROP ROLE {0}", name(MG_USER_PREFIX + user)));
     log(start, "removed user " + user);
+
+    tx(
+        db ->
+            ((SqlDatabase) db)
+                .getJooq()
+                .deleteFrom(USERS_METADATA)
+                .where(USER_NAME.eq(user))
+                .execute());
+    log(start, "removed metadata from user " + user);
+  }
+
+  public void setEnabledUser(String user, Boolean enabled) {
+    long start = System.currentTimeMillis();
+    if (user.equals("admin")) throw new MolgenisException("You cant enable or disable admin");
+    if (user.equals("anonymous")) throw new MolgenisException("You cant enable or disable anonymous");
+    if (!hasUser(user))
+      throw new MolgenisException(
+          (enabled ? "Enabling" : "Disabling")
+              + " user failed: User with name '"
+              + user
+              + "' doesn't exist");
+    tx(
+        db ->
+            ((SqlDatabase) db)
+                .getJooq()
+                .update(USERS_METADATA)
+                .set(USER_ENABLED, enabled)
+                .where(USER_NAME.eq(user))
+                .execute());
+    log(start, (enabled ? "Enabling" : "Disabling") + " user " + user);
   }
 
   public void addRole(String role) {
